refactor[api-server]: [테스트] 테스트 패키지 정리
