--- src/test/java/com/watermelon/server/DocumentConstants.java
@@ -1,10 +1,0 @@
-package com.watermelon.server;
-
-public class DocumentConstants {
-
-    public static final String MY_LINK = "link";
-
-    public static final String PARTS_LINK_LIST = "event/parts/link";
-
-    public static final String SHORTED_LINK = "shortedLink";
-}
