--- src/test/java/com/watermelon/server/admin/service/S3ImageServiceTest.java
@@ -1,50 +1,0 @@
-package com.watermelon.server.admin.service;
-
-import com.amazonaws.auth.AWSCredentials;
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.watermelon.server.admin.exception.S3ImageFormatException;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Profile;
-import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-
-
-@SpringBootTest
-class S3ImageServiceTest {
-
-    private static final Logger log = LoggerFactory.getLogger(S3ImageServiceTest.class);
-    @Autowired
-    private S3ImageService s3ImageService;
-
-    @Test
-    void uploadImage() throws S3ImageFormatException, IOException {
-        String dirName ="image/png";
-        MockMultipartFile image = new MockMultipartFile(
-                "image",
-                "test.jpg",
-                "image/jpeg",
-                dirName.getBytes(StandardCharsets.UTF_8));
-        String imageUrl = s3ImageService.uploadImage(image);
-        System.out.println(imageUrl);
-    }
-
-}
