+++ src/test/java/com/watermelon/server/config/PartsRegistrationConfig.java
@@ -1,0 +1,22 @@
+package com.watermelon.server.config;
+
+import com.watermelon.server.event.parts.domain.Parts;
+import com.watermelon.server.event.parts.repository.PartsRepository;
+import jakarta.annotation.PostConstruct;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.TestConfiguration;
+
+@TestConfiguration
+public class PartsRegistrationConfig {
+
+    @Autowired
+    private PartsRepository partsRepository;
+
+    @PostConstruct
+    public void init() {
+        partsRepository.saveAll(
+                Parts.createAllParts()
+        );
+    }
+
+}
