+++ src/test/java/com/watermelon/server/constants/DocumentConstants.java
@@ -1,0 +1,10 @@
+package com.watermelon.server.constants;
+
+public class DocumentConstants {
+
+    public static final String MY_LINK = "link";
+
+    public static final String PARTS_LINK_LIST = "event/parts/link";
+
+    public static final String SHORTED_LINK = "shortedLink";
+}
