+++ src/test/java/com/watermelon/server/integration/LinkIntegrationTest.java
@@ -1,6 +1,5 @@
 package com.watermelon.server.integration;
 
-import com.watermelon.server.BaseIntegrationTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
